requirements.txt
# requirements.txt
openai
sqlalchemy
psycopg2
pandas
python-dotenv
